Track floor items
We 'x' items to find out what they are now, instead of relying on
the autopickup flag in tiles data, and therefore no longer need
the corpse inscription hack.

We still rely on autopickup to actually pick up items, though.

Disable autoexplore for now so we can see that our manual looting
and exploration is doing the right thing.

Also don't try to pick up items behind tele traps (previously we
only checked whether the item was *on* a tele trap).