Pick up and try-id scrolls
The "Blink to where?" message appears in the targeting mouse mode.
We also sometimes get messages from previous turns in mouse mode,
so previously we ignored all messages received when in mouse mode.
Now, we drop messages whose turn counter is less than the turn
counter of an earlier message.